% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{MappingCalc-package}
\alias{MappingCalc}
\alias{MappingCalc-package}
\title{MappingCalc: Mapping Calculator for EQ-5D Utility Scores}
\description{
Provides a 'shiny' web application to map scores from clinical instruments (PANSS, SQLS, WHODAS 2.0, PHQ-8, EQ-5D-5L) to preference-based EQ-5D-5L health utility values using validated regression-based and beta-mixture mapping algorithms developed from Singapore population studies. Intended for use in health economic evaluations and cost-utility analyses. Methods are based on: Abdin et al. (2019) \doi{10.1007/s11136-018-2037-7}, Seow et al. (2023) \doi{10.1080/14737167.2023.2215430}, Abdin et al. (2021) \doi{10.1186/s12888-021-03463-0}, Abdin et al. (2024) \doi{10.1080/14737167.2024.2376100}.
}
\author{
\strong{Maintainer}: Edimansyah Abdin \email{edimansyah.bin.abdin@gmail.com} (\href{https://orcid.org/0000-0002-1016-3298}{ORCID})

}
