% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqls_utility_score}
\alias{sqls_utility_score}
\title{Estimate EQ-5D-5L Utility Score from SQLS Subscales}
\usage{
sqls_utility_score(psychosocial, motivation, symptoms, age, gender)
}
\arguments{
\item{psychosocial}{Numeric. SQLS Psychosocial subscale score (range: 0--100).}

\item{motivation}{Numeric. SQLS Motivation subscale score (range: 0--100).}

\item{symptoms}{Numeric. SQLS Symptoms subscale score (range: 0--100).}

\item{age}{Numeric. Patient age in years.}

\item{gender}{Numeric. Gender code: 1 = female, 0 = male.}
}
\value{
Numeric scalar. EQ-5D-5L utility value rounded to 3 decimal places.
}
\description{
Predicts the EQ-5D-5L utility index value using an OLS
regression model from Seow et al. (2023).
}
\examples{
sqls_utility_score(50, 50, 50, 40, 1)

}
\references{
Seow E, Abdin E, Subramaniam M, Chong SA (2023). Mapping the schizophrenia
quality of life scale to EQ-5D, HUI3 and SF-6D utility scores in patients
with schizophrenia. \emph{Expert Review of Pharmacoeconomics and Outcomes
Research}, 23(7), 813--821. \doi{10.1080/14737167.2023.2215430}
}
