% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whodas_total_utility_comparison}
\alias{whodas_total_utility_comparison}
\title{Compare WHODAS-Predicted Utility to Singapore Population Mean}
\usage{
whodas_total_utility_comparison(whodas_scores)
}
\arguments{
\item{whodas_scores}{Numeric. WHODAS 2.0 total score (range: 0--48).}
}
\value{
Character. \code{"above"} if predicted utility > 0.95,
\code{"below"} otherwise.
}
\description{
Returns whether the predicted EQ-5D-5L utility is above or
below the Singapore population mean of 0.95.
}
\examples{
whodas_total_utility_comparison(10)

}
