% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whodas_total_utility_score}
\alias{whodas_total_utility_score}
\title{Estimate EQ-5D-5L Utility Score from WHODAS 2.0 Total Score}
\usage{
whodas_total_utility_score(whodas_scores)
}
\arguments{
\item{whodas_scores}{Numeric. WHODAS 2.0 total score (range: 0--48).}
}
\value{
Numeric scalar. EQ-5D-5L utility value rounded to 3 decimal places,
floored at a minimum of -0.584.
}
\description{
Predicts the EQ-5D-5L utility index value from the WHODAS 2.0
12-item total score using a robust regression model from Abdin et al. (2024).
Predicted values below -0.584 are floored at -0.584.
}
\examples{
whodas_total_utility_score(10)
whodas_total_utility_score(48)

}
\references{
Abdin E et al. (2024). Mapping the World Health Organization Disability
Assessment Schedule 2.0 to the EQ-5D-5L in patients with mental disorders.
\emph{Expert Review of Pharmacoeconomics and Outcomes Research}.
\doi{10.1080/14737167.2024.2376100}
}
